+++ main.js
@@ -3,8 +3,11 @@ const express = require('express');
 const routes = require('./index.routes.js');
 const swaggerUi = require('swagger-ui-express');
 const swaggerDocument = require('./swagger.json');
+const bodyParser = require('body-parser');
+const jsonParser = bodyParser.json();
 
 const app = express();
+app.use(jsonParser)
 app.use(routes);
 
 app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerDocument));
