+++ src/authorization/route.signUp.js
@@ -1,10 +1,8 @@
 const express = require('express');
 const routes = express.Router();
 const { controllerSignUp } = require('./controllerSignUp');
-const { bodyValidator } = require('./validate.inputData');
-const bodyParser = require('body-parser');
-const jsonParser = bodyParser.json();
+const { signUpValidator } = require('./validate.inputData');
 
-routes.post('/v1/auth/sign-up', jsonParser, bodyValidator, controllerSignUp);
+routes.post('/v1/auth/sign-up', signUpValidator, controllerSignUp);
 
 module.exports = routes
