+++ src/authorization/validate.schema.js
@@ -1,9 +1,9 @@
 const Joi = require('joi');
 
-const signupSchema = Joi.object({
+const signUpSchema = Joi.object({
     name: Joi.string().alphanum().min(3).max(30).required(),
     email: Joi.string().email().required(),
     password: Joi.string().min(6).required()
 });
 
-module.exports = {signupSchema};
+module.exports = {signUpSchema};
