+++ src/crypto/tests/cipherHash.test.js
